feat(rooter): set rooter management in the app
