+++ src/App.css
@@ -1,0 +1,3 @@
+.app {
+    height: 100vh;
+}
